+++ lib/Controller/AbstractCheckoutController.php
@@ -16,6 +16,7 @@ use Payum\Core\Request\GetHumanStatus;
 
 use Vankosoft\PaymentBundle\Component\OrderFactory;
 use Vankosoft\PaymentBundle\Model\Order;
+use Vankosoft\PaymentBundle\Model\Interfaces\PaymentInterface;
 use Vankosoft\PaymentBundle\EventSubscriber\Event\SubscriptionsPaymentDoneEvent;
 
 abstract class AbstractCheckoutController extends AbstractController
@@ -103,45 +104,31 @@ abstract class AbstractCheckoutController extends AbstractController
     
     protected function paymentSuccess( Request $request, $paymentStatus ): Response
     {
-        $storage    = $this->payum->getStorage( $this->paymentClass );
-        $payment    = $paymentStatus->getFirstModel();
-        //$this->debugObject( $payment );
-        
-        $payment->getOrder()->setStatus( Order::STATUS_PAID_ORDER );
-        $storage->update( $payment );
-        $request->getSession()->remove( 'vs_payment_basket_id' );
-        
+        $payment        = $this->_setPaymentSuccess( $paymentStatus );
         $subscriptions  = $payment->getOrder()->getSubscriptions();
         $hasPricingPlan = ! empty( $subscriptions );
+        $response       = null;
+        $request->getSession()->remove( OrderFactory::SESSION_BASKET_KEY );
         
         if ( $hasPricingPlan ) {
-            $this->eventDispatcher->dispatch(
-                new SubscriptionsPaymentDoneEvent( $subscriptions ),
-                SubscriptionsPaymentDoneEvent::NAME
-            );
-            
-            if ( $this->routeRedirectOnPricingPlanDone ) {
-                $flashMessage   = $this->translator->trans( 'pricing_plan_payment_success', [], 'VSPaymentBundle' );
-                $request->getSession()->getFlashBag()->add( 'notice', $flashMessage );
-                
-                return $this->redirectToRoute( $this->routeRedirectOnPricingPlanDone );
-            }
+            $response   = $this->_setSubscriptionsPaymentDone( $request, $subscriptions );
         }
         
         if ( ! $hasPricingPlan && $this->routeRedirectOnShoppingCartDone ) {
-            $flashMessage   = $this->translator->trans( 'shopping_cart_payment_success', [], 'VSPaymentBundle' );
-            $request->getSession()->getFlashBag()->add( 'notice', $flashMessage );
-            
-            return $this->redirectToRoute( $this->routeRedirectOnShoppingCartDone );
+            $response   = $this->_setShoppingCartPaymentDone( $request );
         }
         
-        return $this->render( '@VSPayment/Pages/Checkout/done.html.twig', [
-            'shoppingCart'                      => $this->orderFactory->getShoppingCart(),
-            'paymentStatus'                     => $paymentStatus,
-            'routeRedirectOnShoppingCartDone'   => $this->routeRedirectOnShoppingCartDone,
-            'routeRedirectOnPricingPlanDone'    => $this->routeRedirectOnPricingPlanDone,
-            'hasPricingPlan'                    => $hasPricingPlan,
-        ]);
+        if ( $response ) {
+            return $response;
+        } else {
+            return $this->render( '@VSPayment/Pages/Checkout/done.html.twig', [
+                'shoppingCart'                      => $this->orderFactory->getShoppingCart(),
+                'paymentStatus'                     => $paymentStatus,
+                'routeRedirectOnShoppingCartDone'   => $this->routeRedirectOnShoppingCartDone,
+                'routeRedirectOnPricingPlanDone'    => $this->routeRedirectOnPricingPlanDone,
+                'hasPricingPlan'                    => $hasPricingPlan,
+            ]);
+        }
     }
     
     protected function paymentFailed( Request $request, $paymentStatus ): Response
@@ -151,7 +138,7 @@ abstract class AbstractCheckoutController extends AbstractController
         
         $payment->getOrder()->setStatus( Order::STATUS_FAILED_ORDER );
         $storage->update( $payment );
-        $request->getSession()->remove( 'vs_payment_basket_id' );
+        $request->getSession()->remove( OrderFactory::SESSION_BASKET_KEY );
         
         throw new HttpException( 400, $this->getErrorMessage( $paymentStatus->getModel() ) );
         return $this->render( '@VSPayment/Pages/Checkout/done.html.twig', [
@@ -161,6 +148,53 @@ abstract class AbstractCheckoutController extends AbstractController
             'routeRedirectOnPricingPlanDone'    => $this->routeRedirectOnPricingPlanDone,
             'hasPricingPlan'                    => false,
         ]);
+    }
+    
+    protected function _setPaymentSuccess( $paymentStatus ): PaymentInterface
+    {
+        $storage    = $this->payum->getStorage( $this->paymentClass );
+        $payment    = $paymentStatus->getFirstModel();
+        //$this->debugObject( $payment );
+        
+        if ( $this instanceof AbstractCheckoutOfflineController ) {
+            $payment->getOrder()->setStatus( Order::STATUS_PENDING_ORDER );
+        } else {
+            $payment->getOrder()->setStatus( Order::STATUS_PAID_ORDER );
+        }
+        
+        $storage->update( $payment );
+        
+        return  $payment;
+    }
+    
+    protected function _setSubscriptionsPaymentDone( Request $request, $subscriptions ): ?Response
+    {
+        if ( $this instanceof AbstractCheckoutOfflineController ) {
+            $flashMessage   = $this->translator->trans( 'vs_payment.template.pricing_plan_payment_success', [], 'VSPaymentBundle' );
+        } else {
+            $this->eventDispatcher->dispatch(
+                new SubscriptionsPaymentDoneEvent( $subscriptions ),
+                SubscriptionsPaymentDoneEvent::NAME
+            );
+            
+            $flashMessage   = $this->translator->trans( 'vs_payment.template.pricing_plan_payment_success', [], 'VSPaymentBundle' );
+        }
+        
+        if ( $this->routeRedirectOnPricingPlanDone ) {
+            $request->getSession()->getFlashBag()->add( 'notice', $flashMessage );
+            
+            return $this->redirectToRoute( $this->routeRedirectOnPricingPlanDone );
+        }
+        
+        return null;
+    }
+    
+    protected function _setShoppingCartPaymentDone( Request $request ): ?Response
+    {
+        $flashMessage   = $this->translator->trans( 'vs_payment.template.shopping_cart_payment_success', [], 'VSPaymentBundle' );
+        $request->getSession()->getFlashBag()->add( 'notice', $flashMessage );
+        
+        return $this->redirectToRoute( $this->routeRedirectOnShoppingCartDone );
     }
     
     protected function getErrorMessage( $details )