+++ lib/Model/PricingPlan.php
@@ -57,9 +57,6 @@ class PricingPlan implements PricingPlanInterface, Comparable
     /** @var Collection|PricingPlanSubscriptionInterface[] */
     protected $subscriptions;
     
-    /** @var bool */
-    protected $recurringPayment = false;
-    
     public function __construct()
     {
         $this->subscriptions    = new ArrayCollection();
@@ -141,7 +138,7 @@ class PricingPlan implements PricingPlanInterface, Comparable
         return $this;
     }
     
-    public function getPaidService(): PayedServiceSubscriptionPeriodInterface
+    public function getPaidService(): ?PayedServiceSubscriptionPeriodInterface
     {
         return $this->paidService;
     }
@@ -222,26 +219,26 @@ class PricingPlan implements PricingPlanInterface, Comparable
         return $this;
     }
     
-    public function isRecurringPayment(): bool
-    {
-        return $this->recurringPayment;
-    }
-    
-    /**
-     * @param bool
-     */
-    public function setRecurringPayment( ?bool $recurringPayment ): PricingPlanInterface
-    {
-        $this->recurringPayment = (bool) $recurringPayment;
-        
-        return $this;
-    }
-    
-    public function getSubscriptionCode(): ?string
+    public function getServiceCode(): ?string
     {
         return $this->paidService->getPayedService()->getSubscriptionCode();
     }
     
+    public function getPeriodCode(): ?string
+    {
+        return $this->paidService->getPaidServicePeriodCode();
+    }
+    
+    public function getSubscriptionCode(): ?string
+    {
+        return $this->getServiceCode() . '-' . $this->getPeriodCode();
+    }
+    
+    /**
+     * @TODO Need to be removed. Use Compare Method.
+     * 
+     * @return int|NULL
+     */
     public function getSubscriptionPriority(): ?int
     {
         return $this->paidService->getPayedService()->getSubscriptionPriority();
@@ -252,6 +249,9 @@ class PricingPlan implements PricingPlanInterface, Comparable
         $period = null;
         
         switch( $this->paidService->getSubscriptionPeriod() ) {
+            case SubscriptionPeriod::SUBSCRIPTION_PERIOD_UNLIMITED:
+                $period = new \DateInterval( 'P1000Y' );
+                break;
             case SubscriptionPeriod::SUBSCRIPTION_PERIOD_YEAR:
                 $period = new \DateInterval( 'P1Y' );
                 break;
@@ -286,7 +286,7 @@ class PricingPlan implements PricingPlanInterface, Comparable
      */
     public function compareTo( $other ): int
     {
-        if ( $this->getSubscriptionCode() != $other->getSubscriptionCode() ) {
+        if ( $this->getServiceCode() != $other->getServiceCode() ) {
             throw new \Exception( 'These Pricing Plans are Not Comparable !!!' );
         }